Almost builds with CMake in a cross-platform way
The hard-coded Android shared memory stuff that I hacked in before
needs to be abstracted out so that different shared memory implementations
can coexist. Currently it won't build for Linux with CMake since it
directly uses ashmem* functions, and it won't build for Android
because the cutils library isn't linked in yet.